v1 (#19)
Update router to be [httpreemux](https://github.com/dimfeld/httptreemux) to expand the path options. The routing rules are relaxed so that a single path segment may be a wildcard in one route and a static token in another. This router is slightly slower than the one removed, but the overhead should be negligible. See benchmarks [here](https://github.com/julienschmidt/go-http-routing-benchmark).

This package is now bumped to v1 because there is a public API change:
- `func SetDetails(r *http.Request, path string, params httprouter.Params) *http.Request` becomes `func SetDetails(r *http.Request, path string, params map[string]string) *http.Request`
- `func GetDetails` is removed from the public API
- `type Details struct` is removed from the public API